flatten hierarchy
